requirements.txt
discord==1.0.1
python-dotenv==0.15.0
requests==2.25.1

aiohttp==3.7.4.post0
NBT==1.5.0

seaborn==0.11.1
matplotlib==3.3.4
numpy==1.20.1
